Sequence opt 2

+++ rasterize_script.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 from __functions import *
 
 
@@ -18,7 +16,7 @@ def create_reference_raster(gpd_frame, reference_gt):
     new_dataset = rasterio.open('temp/reference_raster.tif', 'w', driver='GTiff',
                                 height=arr.shape[0], width=arr.shape[1],
                                 count=1,
-                                dtype=np.dtype(np.int),
+                                dtype=np.dtype(int),
                                 crs=crs,
                                 transform=transform)
     new_dataset.write(arr, 1)
